Major rework of tables to convert into datatables
