+++ babylab/__about__.py
@@ -2,4 +2,4 @@
 About the package.
 """
 
-__version__ = "0.1.7"  # no cov
+__version__ = "0.1.8"  # no cov
