+++ babylab/app/static/styles/style.css
@@ -166,11 +166,16 @@ i {
     text-align: center;
 }
 
+.tbl-record {
+    margin: 1em;
+}
+
 .card-body {
     float: left;
     font-family: "Ciutadella Regular";
     margin: 1em;
     padding: 0%;
+    background-color: white;
 }
 
 .card-img-top {
@@ -203,6 +208,7 @@ i {
     text-wrap: pretty;
     vertical-align: middle;
     word-break: normal;
+    background-color: white;
 }
 
 .tdbreak {
@@ -234,6 +240,7 @@ td {
     background-color: lightgrey;
 }
 
+.dtsp-name,
 .option {
     font-family: "Ciutadella Regular";
 }
@@ -261,6 +268,7 @@ button,
     font-family: "Ciutadella Medium";
 }
 
+
 b,
 h1,
 h2,
@@ -270,6 +278,26 @@ h5,
 h6,
 th {
     font-family: "Ciutadella Semibold";
+}
+
+.dtsp-panes,
+.d-md-flex,
+.mt,
+.dt-container,
+.dt-layout-row,
+.dtsp-panesContainer {
+    padding: 0.5em 2em 0.5em 2em;
+    font-size: medium;
+}
+
+
+.dt-layout-row {
+    border-radius: 1%;
+}
+
+
+.dtsp-title {
+    font-weight: bold;
 }
 
 #indexCardLogin,
