+++ babylab/app/templates/ppt_all.html
@@ -11,15 +11,10 @@
 </header>
 
 <body>
-  <link href="https://cdn.datatables.net/v/dt/dt-2.2.1/b-3.2.1/sp-2.3.3/sl-3.0.0/datatables.min.css" rel="stylesheet">
-  <form method="post">
-    <div class="row">
-      <div class="col-3">
-        <h3>Search for a participant</h3>
-      </div>
-    </div>
-    <br>
-  </form>
+  <link href="https://cdn.datatables.net/v/bs5/jq-3.7.0/dt-2.2.1/b-3.2.1/fh-4.0.1/sp-2.3.3/sl-3.0.0/datatables.min.css"
+    rel="stylesheet">
+  <h3>Search for a participant</h3>
+  <br>
   <a href="{{ url_for('ppt_new') }}">
     <button type="button" class="btn btn-success">
       <i class="fa fa-plus"></i>&nbsp;&nbsp;New
@@ -43,20 +38,22 @@
     </div>
   </div>
   <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
-  <script src="https://cdn.datatables.net/v/dt/dt-2.2.1/b-3.2.1/sp-2.3.3/sl-3.0.0/datatables.min.js"></script>
+  <script
+    src="https://cdn.datatables.net/v/bs5/jq-3.7.0/dt-2.2.1/b-3.2.1/fh-4.0.1/sp-2.3.3/sl-3.0.0/datatables.min.js"></script>
+  <script type="text/javascript" src="{{ url_for('static', filename='js/tables.js') }}"></script>
   <script>
-    let table = new DataTable(
-      '#ppttable',
-      {
-        layout: {
-          topStart: {
-            searchPanes: {
-              columns: [3, 5, 6],
-            },
+    let table = new dt("#ppttable", searchCols = [3, 5, 6], hideCols = [7, 8, 9, 10, 11, 12, 13])
 
-          }
-        },
-      });
+    table.on('click', 'td', function (e) {
+      let tr = e.target.closest('tr');
+      let row = table.row(tr);
+      if (row.child.isShown()) {
+        row.child.hide();
+      }
+      else {
+        row.child(format_ppt(row.data())).show();
+      }
+    });
   </script>
 </body>
 {% endblock content %}
