+++ babylab/src/utils.py
@@ -3,10 +3,8 @@ Util functions for the app.
 """
 
 import os
-from collections import OrderedDict
 import datetime
 import shutil
-import pandas as pd
 from pandas import DataFrame
 from flask import flash, render_template
 from babylab.src import api
@@ -225,50 +223,6 @@ def replace_labels(x: DataFrame | dict, data_dict: dict) -> DataFrame:
     return None
 
 
-def get_participants_table(
-    records: api.Records, data_dict: dict, n: int = None
-) -> DataFrame:
-    """Get participants table
-
-    Args:
-        records (api.Records): REDCap records, as returned by ``api.Records``.
-        data_dict (dict, optional): Data dictionary as returned by ``api.get_data_dictionary``.
-        n (int, optional): Number of records to show. Defaults to None (all records are shown).
-
-    Returns:
-        DataFrame: Table of partcicipants.
-    """  # pylint: disable=line-too-long
-    cols = [
-        "name",
-        "age_now_months",
-        "age_now_days",
-        "sex",
-        "source",
-        "date_created",
-        "date_updated",
-    ]
-    if not records.participants.records:
-        return DataFrame([], columns=cols)
-
-    new_age_months = []
-    new_age_days = []
-    for _, v in records.participants.records.items():
-        age = api.get_age(
-            birth_date=api.get_birth_date(
-                age=f"{v.data['age_now_months']}:{v.data['age_now_days']}"
-            )
-        )
-        new_age_months.append(int(age[0]))
-        new_age_days.append(int(age[1]))
-
-    df = records.participants.to_df()
-    df["age_now_months"] = new_age_months
-    df["age_now_days"] = new_age_days
-    if n:
-        df = df.tail(n)
-    return replace_labels(df, data_dict)
-
-
 def get_age_timestamp(
     apt_records: dict, ppt_records: dict, timestamp: str = "date"
 ) -> tuple[str, str]:
@@ -310,6 +264,74 @@ def get_age_timestamp(
         months_new.append(int(age_now[0]))
         days_new.append(int(age_now[1]))
     return months_new, days_new
+
+
+def get_participants_table(
+    records: api.Records, data_dict: dict, n: int = None
+) -> DataFrame:
+    """Get participants table
+
+    Args:
+        records (api.Records): REDCap records, as returned by ``api.Records``.
+        data_dict (dict, optional): Data dictionary as returned by ``api.get_data_dictionary``.
+        n (int, optional): Number of records to show. Defaults to None (all records are shown).
+
+    Returns:
+        DataFrame: Table of partcicipants.
+    """  # pylint: disable=line-too-long
+    cols = [
+        "record_id",
+        "date_created",
+        "date_updated",
+        "source",
+        "name",
+        "age_now_months",
+        "age_now_days",
+        "days_since_last_appointment",
+        "sex",
+        "twin",
+        "parent1_name",
+        "parent1_surname",
+        "email1",
+        "phone1",
+        "parent2_name",
+        "parent2_surname",
+        "email2",
+        "phone2",
+        "address",
+        "city",
+        "postcode",
+        "birth_type",
+        "gest_weeks",
+        "birth_weight",
+        "head_circumference",
+        "apgar1",
+        "apgar2",
+        "apgar3",
+        "hearing",
+        "diagnoses",
+        "comments",
+    ]
+    if not records.participants.records:
+        return DataFrame([], columns=cols)
+
+    new_age_months = []
+    new_age_days = []
+    for _, v in records.participants.records.items():
+        age = api.get_age(
+            birth_date=api.get_birth_date(
+                age=f"{v.data['age_now_months']}:{v.data['age_now_days']}"
+            )
+        )
+        new_age_months.append(int(age[0]))
+        new_age_days.append(int(age[1]))
+
+    df = records.participants.to_df()
+    df["age_now_months"] = new_age_months
+    df["age_now_days"] = new_age_days
+    if n:
+        df = df.tail(n)
+    return replace_labels(df, data_dict)
 
 
 def get_appointments_table(
@@ -458,443 +480,6 @@ def count_col(
             if idx > 0:
                 counts[k] = v + list(counts.values())[idx - 1]
     return counts
-
-
-def prepare_dashboard(
-    records: api.Records = None, data_dict: dict = None, **kwargs
-) -> dict:
-    """Prepare data for dashboard.
-
-    Args:
-        records (api.Records): REDCap records, as returned by ``api.Records``.
-        data_dict (dict, optional): Data dictionary as returned by ``api.get_data_dictionary``. Defaults to None.
-        **kwargs: Extra arguments passed to ``get_participants_table``, ``get_appointments_table``, and ``get_questionnaires_table``
-
-    Returns:
-        dict: Parameters for the dashboard endpoint.
-    """  # pylint: disable=line-too-long
-    ppts = get_participants_table(records, data_dict=data_dict, **kwargs)
-    apts = get_appointments_table(records, data_dict=data_dict, **kwargs)
-    quest = get_questionnaires_table(records, data_dict=data_dict, **kwargs)
-    ppts["age_days"] = round(
-        ppts["age_now_days"] + (ppts["age_now_months"] * 30.437), None
-    ).astype(int)
-    age_bins = list(range(0, max(ppts["age_days"]), 15))
-    labels = [f"{int(a // 30)}:{int(a % 30)}" for a in age_bins]
-    ppts["age_days_binned"] = pd.cut(
-        ppts["age_days"], bins=age_bins, labels=labels[:-1]
-    )
-
-    variables = {
-        "age_dist": count_col(ppts, "age_days_binned"),
-        "sex_dist": count_col(ppts, "sex", values_sort=True),
-        "source_dist": count_col(ppts, "source", values_sort=True),
-        "ppts_date_created": count_col(ppts, "date_created", cumulative=True),
-        "apts_date_created": count_col(apts, "date_created", cumulative=True),
-        "status_dist": count_col(apts, "status", values_sort=True),
-        "lang1_dist": count_col(quest, "lang1", values_sort=True, missing_label="None"),
-        "lang2_dist": count_col(quest, "lang2", values_sort=True, missing_label="None"),
-    }
-
-    return {
-        "n_ppts": ppts.shape[0],
-        "n_apts": apts.shape[0],
-        "age_dist_labels": list(variables["age_dist"].keys()),
-        "age_dist_values": list(variables["age_dist"].values()),
-        "sex_dist_labels": list(variables["sex_dist"].keys()),
-        "sex_dist_values": list(variables["sex_dist"].values()),
-        "source_dist_labels": list(variables["source_dist"].keys()),
-        "source_dist_values": list(variables["source_dist"].values()),
-        "ppts_date_created_labels": list(variables["ppts_date_created"].keys()),
-        "ppts_date_created_values": list(variables["ppts_date_created"].values()),
-        "apts_date_created_labels": list(variables["apts_date_created"].keys()),
-        "apts_date_created_values": list(variables["apts_date_created"].values()),
-        "status_dist_labels": list(variables["status_dist"].keys()),
-        "status_dist_values": list(variables["status_dist"].values()),
-        "lang1_dist_labels": list(variables["lang1_dist"].keys())[:24],
-        "lang1_dist_values": list(variables["lang1_dist"].values())[:24],
-        "lang2_dist_labels": list(variables["lang2_dist"].keys())[:24],
-        "lang2_dist_values": list(variables["lang2_dist"].values())[:24],
-    }
-
-
-def prepare_participants(records: api.Records, data_dict: dict, **kwargs) -> dict:
-    """Prepare data for participants page.
-
-    Args:
-        records (api.Records): REDCap records, as returned by ``api.Records``.
-        data_dict (dict): Data dictionary as returned by ``api.get_data_dictionary``.
-        **kwargs: Extra arguments passed to ``get_participants_table``.
-
-    Returns:
-        dict: Parameters for the participants endpoint.
-    """  # pylint: disable=line-too-long
-    df = get_participants_table(records, data_dict=data_dict, **kwargs)
-    classes = "table table-hover table-responsive"
-    df["record_id"] = [format_ppt_id(i) for i in df.index]
-    df.index = df.index.astype(int)
-    df = df.sort_index(ascending=False)
-    df["modify_button"] = [format_modify_button(p) for p in df.index]
-    df = df[
-        [
-            "record_id",
-            "name",
-            "age_now_months",
-            "age_now_days",
-            "sex",
-            "source",
-            "date_created",
-            "date_updated",
-            "modify_button",
-        ]
-    ]
-    df = df.rename(
-        columns={
-            "record_id": "Participant",
-            "name": "Name",
-            "age_now_months": "Age (months)",
-            "age_now_days": "Age (days)",
-            "sex": "Sex",
-            "source": "Source",
-            "date_created": "Added on",
-            "date_updated": "Last updated",
-            "modify_button": "",
-        }
-    )
-    return {
-        "table": df.to_html(
-            classes=f'{classes}" id = "ppttable',
-            escape=False,
-            justify="left",
-            index=False,
-            bold_rows=True,
-        )
-    }
-
-
-def prepare_record_id(
-    records: api.Records, data_dict: dict, ppt_id: str, **kwargs
-) -> dict:
-    """Prepare record ID page.
-
-    Args:
-        records (api.Records): REDCap records, as returned by ``api.Records``.
-        ppt_id (str, optional): Participant ID. Defaults to None.
-        **kwargs: Extra arguments passed to ``get_participants_table``, ``get_appointments_table``, and ``get_questionnaires_table``
-
-    Returns:
-        dict: Parameters for the participants endpoint.
-    """  # pylint: disable=line-too-long
-    data = records.participants.records[ppt_id].data
-    for k, v in data.items():
-        kdict = "participant_" + k
-        if kdict in data_dict:
-            data[k] = data_dict[kdict][v] if v else ""
-    data["age_now_months"] = (
-        str(data["age_now_months"]) if data["age_now_months"] else ""
-    )
-    data["age_now_days"] = str(data["age_now_days"]) if data["age_now_days"] else ""
-    data["parent1"] = data["parent1_name"] + " " + data["parent1_surname"]
-    data["parent2"] = data["parent2_name"] + " " + data["parent2_surname"]
-
-    classes = "table table-hover table-responsive"
-
-    # prepare participants table
-    df_appt = get_appointments_table(
-        records, data_dict=data_dict, ppt_id=ppt_id, **kwargs
-    )
-    df_appt["record_id"] = [format_ppt_id(i) for i in df_appt.index]
-    df_appt["appointment_id"] = [format_apt_id(i) for i in df_appt["appointment_id"]]
-    df_appt = df_appt.sort_values(by="date", ascending=False)
-    df_appt = df_appt[
-        [
-            "record_id",
-            "appointment_id",
-            "study",
-            "date",
-            "date_created",
-            "date_updated",
-            "taxi_address",
-            "taxi_isbooked",
-            "status",
-        ]
-    ]
-    df_appt = df_appt.rename(
-        columns={
-            "record_id": "Participant",
-            "appointment_id": "Appointment",
-            "study": "Study",
-            "date": "Date",
-            "date_created": "Made on the",
-            "date_updated": "Last update",
-            "taxi_address": "Taxi address",
-            "taxi_isbooked": "Taxi booked",
-            "status": "Status",
-        }
-    )
-    table_appt = df_appt.to_html(
-        classes=classes,
-        escape=False,
-        justify="left",
-        index=False,
-        bold_rows=True,
-    )
-
-    # prepare language questionnaires table
-    df_quest = get_questionnaires_table(records, data_dict=data_dict, ppt_id=ppt_id)
-    df_quest["questionnaire_id"] = [
-        format_que_id(p, q)
-        for p, q in zip(df_quest.index, df_quest["questionnaire_id"])
-    ]
-    df_quest["record_id"] = [format_ppt_id(i) for i in df_quest.index]
-    df_quest = df_quest[
-        [
-            "questionnaire_id",
-            "record_id",
-            "lang1",
-            "lang1_exp",
-            "lang2",
-            "lang2_exp",
-            "lang3",
-            "lang3_exp",
-            "lang4",
-            "lang4_exp",
-            "date_created",
-            "date_updated",
-        ]
-    ]
-    df_quest = df_quest.sort_values("date_created", ascending=False)
-    df_quest = df_quest.rename(
-        columns={
-            "record_id": "ID",
-            "questionnaire_id": "Questionnaire",
-            "date_updated": "Last updated",
-            "date_created": "Created on the:",
-            "lang1": "L1",
-            "lang1_exp": "%",
-            "lang2": "L2",
-            "lang2_exp": "%",
-            "lang3": "L3",
-            "lang3_exp": "%",
-            "lang4": "L4",
-            "lang4_exp": "%",
-        }
-    )
-
-    table_quest = df_quest.to_html(
-        classes=classes,
-        escape=False,
-        justify="left",
-        index=False,
-        bold_rows=True,
-    )
-
-    return {
-        "data": data,
-        "table_appointments": table_appt,
-        "table_questionnaires": table_quest,
-    }
-
-
-def prepare_appointments(
-    records: api.Records, data_dict: dict = None, study: str = None, **kwargs
-):
-    """Prepare record ID page.
-
-    Args:
-        records (api.Records): REDCap records, as returned by ``api.Records``.
-        data_dict (dict): Data dictionary as returned by ``api.get_data_dictionary``.
-        study (str, optional): Study to filter for. Defaults to None.
-        **kwargs: Extra arguments passed to ``get_participants_table``, ``get_appointments_table``, and ``get_questionnaires_table``
-
-    Returns:
-        dict: Parameters for the participants endpoint.
-    """  # pylint: disable=line-too-long
-    df = get_appointments_table(records, data_dict=data_dict, study=study, **kwargs)
-    classes = "table table-hover table-responsive"
-    df["record_id"] = [format_ppt_id(i) for i in df.index]
-    df["modify_button"] = [
-        format_modify_button(p, a) for p, a in zip(df.index, df["appointment_id"])
-    ]
-    df["appointment_id"] = [format_apt_id(i) for i in df["appointment_id"]]
-    df["status"] = [format_status(s) for s in df["status"]]
-
-    df = df[
-        [
-            "appointment_id",
-            "record_id",
-            "study",
-            "status",
-            "date",
-            "date_created",
-            "date_updated",
-            "taxi_address",
-            "taxi_isbooked",
-            "modify_button",
-        ]
-    ]
-    df = df.sort_values("date_updated", ascending=False)
-
-    df = df.rename(
-        columns={
-            "appointment_id": "Appointment",
-            "record_id": "Participant",
-            "study": "Study",
-            "status": "Appointment status",
-            "date": "Date",
-            "date_created": "Made on the",
-            "date_updated": "Last updated",
-            "taxi_address": "Taxi address",
-            "taxi_isbooked": "Taxi booked",
-            "modify_button": "",
-        }
-    )
-
-    table = df.to_html(
-        classes=f'{classes}" id = "apttable',
-        escape=False,
-        justify="left",
-        index=False,
-        bold_rows=True,
-    )
-
-    return {"table": table}
-
-
-def prepare_questionnaires(records: api.Records, data_dict: dict, **kwargs):
-    """Prepare appointments page.
-
-    Args:
-        records (api.Records): REDCap records, as returned by ``api.Records``.
-        data_dict (dict): Data dictionary as returned by ``api.get_data_dictionary``.
-        **kwargs: Extra arguments passed to ``get_participants_table``, ``get_appointments_table``, and ``get_questionnaires_table``
-
-    Returns:
-        dict: Parameters for the participants endpoint.
-    """  # pylint: disable=line-too-long
-    df = get_questionnaires_table(records, data_dict=data_dict, **kwargs)
-    classes = "table table-hover"
-    df["modify_button"] = [
-        format_modify_button(p, ques_id=q)  # pylint: disable=line-too-long
-        for p, q in zip(df.index, df["questionnaire_id"])
-    ]
-    df["questionnaire_id"] = [
-        format_que_id(p, q) for p, q in zip(df["questionnaire_id"], df.index)
-    ]
-    df["record_id"] = [format_ppt_id(i) for i in df.index]
-    df = df[
-        [
-            "questionnaire_id",
-            "record_id",
-            "isestimated",
-            "lang1",
-            "lang1_exp",
-            "lang2",
-            "lang2_exp",
-            "lang3",
-            "lang3_exp",
-            "lang4",
-            "lang4_exp",
-            "date_updated",
-            "date_created",
-            "modify_button",
-        ]
-    ]
-    df = df.sort_values("date_created", ascending=False)
-    df = df.rename(
-        columns={
-            "record_id": "Participant",
-            "questionnaire_id": "Questionnaire",
-            "isestimated": "Status",
-            "date_updated": "Last updated",
-            "date_created": "Added on the:",
-            "lang1": "L1",
-            "lang1_exp": "%",
-            "lang2": "L2",
-            "lang2_exp": "%",
-            "lang3": "L3",
-            "lang3_exp": "%",
-            "lang4": "L4",
-            "lang4_exp": "%",
-            "modify_button": "",
-        }
-    )
-
-    table = df.to_html(
-        classes=f'{classes}" id = "quetable',
-        escape=False,
-        justify="left",
-        index=False,
-        bold_rows=True,
-    )
-
-    return {"table": table}
-
-
-def prepare_studies(records: api.Records, data_dict: dict, study: str = None):
-    """Prepare appointments page.
-
-    Args:
-        records (api.Records): REDCap records, as returned by ``api.Records``.
-        data_dict (dict): Data dictionary as returned by ``api.get_data_dictionary``.
-        study (str, optional): Study to filter for. Defaults to None.
-
-    Returns:
-        dict: Parameters for the participants endpoint.
-    """  # pylint: disable=line-too-long
-    df = get_appointments_table(records, data_dict=data_dict, study=study)
-    classes = "table table-hover table-responsives"
-    df["appointment_id"] = [format_apt_id(i) for i in df["appointment_id"]]
-    df["record_id"] = [format_ppt_id(i) for i in df.index]
-    df = df[
-        [
-            "appointment_id",
-            "record_id",
-            "study",
-            "date",
-            "date_created",
-            "date_updated",
-            "taxi_address",
-            "taxi_isbooked",
-            "status",
-        ]
-    ]
-    df = df.sort_values("date", ascending=False)
-
-    df = df.rename(
-        columns={
-            "appointment_id": "Appointment",
-            "record_id": "Participant",
-            "study": "Study",
-            "date": "Date",
-            "date_created": "Made on the",
-            "date_updated": "Last updated",
-            "taxi_address": "Taxi address",
-            "taxi_isbooked": "Taxi booked",
-            "status": "Appointment status",
-        }
-    )
-
-    table = df.to_html(
-        classes=f'{classes}" id = "apttable',
-        escape=False,
-        justify="left",
-        index=False,
-        bold_rows=True,
-    )
-
-    date = df["Date"].value_counts().to_dict()
-    date = OrderedDict(sorted(date.items()))
-    for idx, (k, v) in enumerate(date.items()):
-        if idx > 0:
-            date[k] = v + list(date.values())[idx - 1]
-
-    return {
-        "n_apts": df.shape[0],
-        "date_labels": list(date.keys()),
-        "date_values": list(date.values()),
-        "table": table,
-    }
 
 
 def clean_tmp(path: str = "tmp"):
